+++ CHANGE_LOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Add ability for widget to be draggable on desktop browser
 - Added warning banner for third party cookies blocked alert
 
 ### Fixed